+++ lib/garner/strategies/keys/caller_strategy.rb
@@ -16,15 +16,19 @@ module Garner
               return rc
             end
 
-            caller.each do |line|
-              split = line.split(":")
-              next unless split.length >= 2
-              path = (Pathname.new(split[0]).realpath.to_s rescue nil)
-              next if path.blank? || path.include?("lib/garner")
-              next unless path.include?("/app/") || path.include?("/spec/")
-              rc[field] = "#{path}:#{split[1]}"
-              break
+            if caller
+              caller.each do |line|
+                next unless line
+                split = line.split(":")
+                next unless split && split.length >= 2
+                path = (Pathname.new(split[0]).realpath.to_s rescue nil)
+                next if (! path) || path.empty? || path.include?("lib/garner")
+                next unless path.include?("/app/") || path.include?("/spec/")
+                rc[field] = "#{path}:#{split[1]}"
+                break
+              end
             end
+
             rc
           end
         end